apply cody vscode config changes without reload (#50695)
The code could be a bit cleaner here, but overall this is a user win, as
well as a code simplification that makes future refactors easier.

For the reviewer: I think we should merge this soon and do code cleanups
as a followup. This will cause a lot of merge conflicts if it lingers.

## Test plan

Sign out of Cody in VS Code. Sign in again. Toggle `cody.debug`. Ensure
that changes take effect and there is no reload needed.