show "reset conversation" icon in view/title menu not react UI (#50964)
The standard place for this kind of button is in the view/title menu,
not in the webview UI. This change also makes it use the standard VS
Code icon for this type of action, instead of a non-VS Code-standard
icon (the reload).

This is similar to how we moved settings to a gear icon in the sme
view/title menu.

<img width="511" alt="image"
src="https://user-images.githubusercontent.com/1976/233556675-c5bdd150-ad09-46d8-980c-9d5ea717d9ff.png">



## Test plan

Click the button after you've started a conversation. Confirm it is
cleared and reset.