[language-platform] Remove unnecessary dependencies in upload service (#50870)
This pull request aims to remove unnecessary dependencies in the upload
service codebase. Upon reviewing the current code, it was found that
some dependencies are no longer being used and only add unnecessary
overhead to the application.

The following dependencies have been identified for removal:

- policyService
- policyMatcher
- rankingBucket
- locker

By removing these dependencies, we can improve the overall performance
and maintainability of the upload service. This change does not affect
any of the existing functionalities of the service.

## Test plan
The changes in this pull request have been thoroughly tested and
reviewed to ensure that they do not cause any unintended consequences or
regressions. If there are any concerns or questions regarding this pull
request, please do not hesitate to raise them in the comments section.
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->