Embeddings: exclude files by size in `embedFiles` (#51021)
Currently, we exclude files by their size when we create the input list
of files. The problem with this is it's pretty distant from the place we
do all the other file skipping, so it makes collecting stats about why
files were skipped more difficult.

This updates the code to maintain the list of files along with their
size so that we can add them to the excluded counts. If we do not list
them, we cannot count them. This has the added bonus that we can keep
track of the _sizes_ of files that were skipped, so we have a full
bookkeeping of how much of the repo was indexed.