Cody: Recent changes recipes message (#51098)
This PR is for issue #51084. There was a console message when no git log
was present for the selected dropdown. This PR implemented the
functionality to return a response as "No recent changes found".

## Test plan

All the manual and integration tests have passed.
Manual testing has been done to check the expected behaviour.