add Cody prompt for unauthd users (#51229)
Right now, any visitor to dotcom can see Cody in the UI. We're making a
separate change to only enable Cody UI on dotcom for users who are
signed in. This change prompts users who are NOT signed in to
signin/sign up in order to get access to Cody.