Cody completions: Add metrics for completion events (#51350)
Closes #51347 

This adds metrics for inline and multiline suggestions so we can know
how often we trigger requests, how often they are presented to our
users, and what the acceptance rate is (note: acceptance rate doesn't
work for multiline suggestions yet as we only allow copy-pasting for
this and don't have a distinct UI action for accepting)

## New events added:

- `CodyVSCodeExtension:completion:started` (`type: 'inline' |
'multiline'`)
- Measure when we start a request to the backend for new suggestions
(and in turn hit the LLM APIs). I wanted to add this so we can calculate
how often these are aborted.
- `CodyVSCodeExtension:completion:suggested` (`type: 'inline' |
'multiline'`)
- Measures when we receive a result from the backend and are displaying
those to the user (Note: VS Code does not always present inline
suggestions to the user, even if this event fires. We'll have to
investigate why that's the case and make sure we don't start a
completion unless they will be shown).
- `CodyVSCodeExtension:completion:accepted` (`type: 'inline'`)
  - Measures how often completions are inserted. 

## Open question: Logging completion on/off (cc @malomarrec)

It's not clear yet how we can measure total number of people that have
suggestions enabled (changing the setting for this is not a discrete
event since it can happen outside of our control). We could include the
state of the feature flag in other Cody related pings but I really think
given the occurence of inline completions, these will swiftly become the
most-frequently-used Cody feature (when enabled). This leads me to
believe that we should calculate DAUs based on the presence of an inline
suggestion event.

## Test plan

- Add a console.log to the logging code so we can visualize logged
events in the console
-
https://user-images.githubusercontent.com/458591/235658549-779b1b4f-0603-49fe-9b47-3ca94a7393ab.mov



<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->