Embeddings: allow max embedding counts to be configurable (#51326)
This adds the `maxCodeEmbeddingsPerRepo` and `maxTextEmbeddingsPerRepo`
config options to our embeddings config. This will allow us to adjust
the max to support larger monorepos without pushing a new patch release.

As part of this change, I also included a small refactor that puts all
`EmbedRepo` options into a struct because the function args were getting
a little overwhelming.