App: new completions provider to forward requests to dotcom (#51426)
The adds a new completions provider for use in the Sourcegraph App that
will send completions requests to sourcegraph.com. This will allow users
of Sourcegraph App to use Cody within App by using a dotcom account and
they will not need to configure App with an LLM key. App users that do
have their own API keys can apply a completions configuration if they
wish communicate directly with the LLM providers.

I chose not to add this provider as a configuration option because it
should only be used for Sourcegraph App and I wanted to avoid any
accidentally configurations using it.