bazel: run backend tests with bazel (#51438)
Migrates our backend tests to use bazel. 

Also removes all the separate go tests in favor of `bazel test` as this
covers everything in ~40 seconds.

Introduces some logic in the backendIntegrationTests function so we can
change the image dependency based on the run type.

## Test plan

Green CI on this PR is sufficient.