Cody completions: Fix prompt size calculation for inline suggestions (#51505)
I noticed an error in the inline completion code. We create an "empty"
completion (so one with no third-party code snippets) to measure the
minimum size of the prompt. However, beacuse of a copy-paste error we
were using the wrong prompt for the size (so we were creating the text
prompt for a multiline completion and used this for inline prompts.
Whoops.

I didn't notice any errors with this which might mean that multiline
prompts >= inline prompts but better fix it 😬

## Test plan

Inline completions still work

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->