search: apply limits in zoekt when doing select:repo (#51566)
Previously we didn't set any limits other than ShardRepoMaxMatchCount=1
if we had filter.Repository. However, the limits are still useful. In
practice it is rare this would make much of a difference except in the
case of large compound shards and low limits. Either way it feels
correct to still pass along the limits.

Test Plan: updated unit tests