llmproxy: Implement SourceSyncer for productsubscription source (#51689)
This implements a periodic full sync on the productsubscription actor
source by listing all the subscriptions from the dotcom API and then
inserting a record for each of the items. This is a bit wasteful as we
keep the same payload for all of them, but it's the simplest we can get
with our current caching. I've also not yet implemented pruning old
keys, that we should follow up on.