bazel: disable the bundlesize check outside of bazel (#51758)
Let's monitor the behavior of this task. If it happens again, we disable
it, and I'll dig further. But I'm sure I saw successful Bazel builds
with the bundlesize cache error. [The Slack
thread](https://sourcegraph.slack.com/archives/C02FLQDD3TQ/p1683634881118919).

This PR also disables the bundlesize check outside of Bazel so that we
run it only once per CI build.