Tracing: clean up uses of opentracing in search code (#51856)
This removes nearly all the dependencies on opentracing from the search
codebase.

The search codebase used the opentracing log types heavily for logging,
tracing, and pretty printing our search jobs. And because our search
jobs all conform to a common interface, it would be difficult to
approach this incrementally. So I just ripped the band-aid off and
converted everything to use the opentelemetry types.

Now the only dependent of the opentracing log types is the observation
package.