Executors: add general dequeue endpoint (#51892)
- Closes https://github.com/sourcegraph/sourcegraph/issues/50613

This PR adds a first implementation of a general dequeue endpoint that
can accept multiple queues.

### Improvements
Due to generics, it's pretty difficult to not break this up into queue
specific components due to the record types handled by each queue. This
means there's still a bunch of reproduced code in the current
implementation.

It currently imlements `ServeHTTP`. Once
https://github.com/sourcegraph/sourcegraph/issues/51656 is added, it
might make sense to turn this into a new interface.

## Demo
### Only batch change jobs

https://github.com/sourcegraph/sourcegraph/assets/2979513/ebe6844c-0fb9-4207-a3ce-f651a1a48631

### Only codeintel jobs

https://github.com/sourcegraph/sourcegraph/assets/2979513/af027eed-6950-4a55-b94c-73847f31ddb8

### Both queue jobs

https://github.com/sourcegraph/sourcegraph/assets/2979513/255c8716-15cb-44e4-9241-f5e28d8d8c7e


## Test plan
- [x] Unit tests
- [x] Demo

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Randell Callahan <piszmogcode@gmail.com>