Observation: add an option to use OTel attrs rather than Opentracing fields (#51954)
This adds an additional `Attrs` field to the `observation.Args` type
that can be used instead of `LogFields`. Before they are actually used,
`LogFields` and `Attrs` will be merged, so right now, using either one
will work fine.

This allows us to incrementally remove more dependencies on the
opentracing log fields rather than having to convert every usage of
`Args.LogFields` at once, which would be intense.

This also includes one migration of `LogFields` to `Attrs` as an
example.