Add temporary default value for completions model (#52025)
This adds a temporary value to fix some instances where the completion
model is not set, as the client will not always send an override for
that anymore soon.

## Test plan

Doesn't fail anymore when completions is not configured. 
I think the proper next step could be to disable the completions
endpoint if not set, or something. Or perhaps a provider specific
default. This is mostly to fix existing instances right now.