Prepare for support for OpenAI based code completions (#52042)
This adds the required backend functionality to support code completions
with OpenAI. It does NOT implement code completions though, as the
client has to craft and parse the queries and responses in a way that
currently seems very specific to Anthropic. If someone is interested,
this would be the next step to feature parity here. After this PR, using
a GPT model like gpt-3.5-turbo can be used to ask for code completions.
It returns a valid response to the VSCode extension, but the contents of
it are not useful so nothing really shows up.

This PR is mostly to prepare the API ahead of the client because the
client can ship any time.