App: Use arguments to pass data locations to sg background process (#52076)
This moves the responsibility of setting the data and cache directories
from the embedded binary to the Tauri shell sent to the app via command
line arguments. The purpose is so that code within the Tauri shell can
interact with the data folders which requires that we alert Tauri to any
files we will work with on the local system in the config. In the future
this will allow things like reseting the database to an empty state.

This does not change the data location when running `sg start app` and
only appends the `sourcegraph(-dev)` folder if the location was not
specified via an argument.

I also updated the bundle identifier to `com.sourcegraph.cody` I expect
we will be renaming the app soon and want to minimize additional changes
in where the data is stored as well as the fact that folders ending in
`.app` are interpreted in macos as an application and is confusing to
see.