app: enable updater in app bundling and generate manifest (#52318)
This enables the updater in during the tauri bundling. In particular a
signature gets generated which we need in our update manifest.
This manifest is used in conjunction with the app updater endpoint on
dotcom. The manifest is consultated to tell clients whether they can
update or not

## Test plan

[Green](https://buildkite.com/sourcegraph/sourcegraph-app-release/builds/1288#01884980-d278-433a-8e37-fb8b979be18a)
ci on `app-release/debug` branch
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->