Worker: use `string` ID instead of `int` for job tracking (#52454)
- Prerequisite for
https://github.com/sourcegraph/sourcegraph/issues/51658, see [thread for
context](https://sourcegraph.slack.com/archives/C04K1NSCK6K/p1684841039315679)

This PR lets the worker keep track of it's running job IDs using a
string ID rather than int IDs. To achieve this, the `Record` interface
is extended with a second method `RecordUID()` which returns a string.

All the types that implement `Record` add a method `RecordUID()` which
simply converts its int value to a string representation. The only
exception to this rule is `executortypes.Job`, which includes the queue
name if it is set. The end goal here is that an executor can process
jobs from multiple queues by retrieving the required context from the
heartbeat IDs without leaking implementation details to the worker
level.

## Test plan
Unit tests.
Looking for input on other ways to validate nothing is broken.
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->