luasandbox: Make modules load on windows (#52760)
The filepath package uses the host systems separator. However, embed.FS
is ALWAYS a unix file system with `/` separators, so on windows, the
luasandbox is currently not able to load the modules correctly. What a
footgun :sadge:

Pulling this out into a separate PR so it's easier to revert if this is
causing issues. I don't think it should cause any merge conflicts with
https://github.com/sourcegraph/sourcegraph/pull/52384

## Test plan

We verified that this fixed an issue on windows.