executors: Handle pod that is never scheduled (#52788)
If a pod cannot be scheduled in a node it will sit in pending state
forever. Eventually the scheduler will delete the pod. I added a check
to look at the deletion timestamp to see if the scheduler deleted the
pod.

When this happens, I want to bubble up a more helpful error instead of
waiting for the context timeout.

## Test plan

Added additional tests.