cmd/gitserver: Move common.RunWith inside server (#52817)
In this commit we move common.RunWith inside the server package and
rename it to make it private.

In #52334, we moved this to the common package because it was required
in #51860. However after iterating further we do not need it any longer
and we move this back into server along with all the other related
symbols.