Embeddings: respect minimum interval for non-completed jobs (#52799)
In #52505 we added a minimum interval between embeddings jobs. It only
checked for completed jobs within the time window, which meant that if a
job was queued, processing, or cancelled, the policy would immediately
attempt to schedule another.

This PR updates the minimum interval check to include all recent jobs,
regardless of state. The semantics are now "we will not schedule new job
if we've already kicked off a job for that repo in the past 24h".