gRPC: enable zoekt separately (#53118)
I merged/rolled out gRPC for Zoekt yesterday, but it ran into
[issues](https://sourcegraph.slack.com/archives/C04HCK4K3DL/p1686100826372909).
This feature-flags the gRPC API for Zoekt separately from the primary
config option so it can be tested separately (and doesn't require a
deployment to enable/disable). In retrospect, I should have rolled it
out this way to start.