App - fix bad version handling on updater handler (#53478)
The App does not encode the url to the updater and because the App
version numbers have a + sign it's being interpreted as a blank space.
This fails the semver comparison and is why app users are not prompted
for auto updates. This will now replace any blank spaces within the
version with a '+'.