App - new mutation to schedule newly added repos for embeddings (#53547)
This adds a new mutation for App only that schedules repos for
embeddings generation. It waits to ensure the repo has been cloned so
that the embeddings job can be successfully created. This is intended to
be used during app setup just after repos are added, previously there
was a race condition between the cloning and scheduling steps.