Embeddings: close pipe with error (#53715)
If we encounter an error while gob encoding the index, we close the pipe
without an error. This causes the upload store to store the index
successfully rather than failing as it should.

This PR modifies the logic so it will close the pipe on error rather
than allowing the upload to succeed. This is a possible explanation for
some index corruption we were seeing.