Settings: make DB calls sequentially (#53803)
We're getting a [concurrent transaction
error](https://sourcegraph.sentry.io/issues/4263654659/?project=6583153&referrer=slack)
from settings because somewhere we're passing a transaction as the
`database.DB` then using it concurrently in this method.

This updates the DB calls to run sequentially. This comes with a perf
penalty, but I'd prefer a small perf impact to
erroring/crashing/silently doing the wrong thing. As a followup, I will
add a DB method that allows you to request settings for multiple
subjects at once. That might not come until after 5.1 though, so I'd at
least like to get this merged and backported.