App: Generate token and write app token file for seamless integration with editor extensions (#53782)
Adds automatic App token generation at startup which can be used by
clients to establish a connection with App.

This allows automatic and seamless connections between App and editor
extensions.

Clients can read `app.json` created by App and use it to make requests
to App's local API.

## How it works

- On startup, App checks if it has already written an `app.json` file
containing a valid token in the App config dir.
- If it's not there or if the token isn't valid, App generates a token
and writes it to `app.json` in the App config dir.
- This takes places after the sign-in link is loaded, which is when the
App user becomes available, because the creation of access tokens
requires a user ID.
- App also writes the localhost API endpoint and the version string in
`app.json`.

## Security considerations

- The token is a local token which is only valid for the local
installation of App. It doesn't grant access to any resources or
services outside of the local environment. It grants the same
permissions as the local App user already has within the App UI.

## Location of app.json

The file is written in the "app config" directory which varies by OS:

- MacOS: `~/Library/Application Support/com.sourcegraph.cody/app.json`


## Contents of app.json

```json
{
  "token": "xxxxx",
  "endpoint": "http://localhost:3080",
  "version": "2023.06.16"
}
```

## Test plan

- Start app
- Check `{APP_CONFIG_DIR}/app.json`
- It should be re-generated if it's deleted