search: populate Stats.Repos and ReposMap from List (#54024)
Not aggregating ReposMap was an oversight which luckily we do not use
yet (but intend on migrating too). Stats.Repos is a field we want to
start using to avoid the need of caching the full list in memory.

Test Plan: updated our fakes to populate Repos and ReposMap which then
exercised the changed code paths.