ci: `percy` defer snapshots upload (#54116)
When running client integration tests, we create a new Percy build on a
first `percySnapshot` call and upload snapshots as soon as they are
taken. If a build fails because of a flake and an auto-retry is
triggered, we create another Percy build and upload all snapshots again.
This is the reason for the rising Percy costs since migrating to Bazel
because the number of snapshots didn't change since then. See [this
Slack
thread](https://sourcegraph.slack.com/archives/C04931KQVRC/p1686911527274019)
for more details.

This PR defers Percy's build creation and snapshots upload until the
test suite completion. No Percy build will be created if tests fail so
that we can keep the auto-retry logic without changes.