guardrails: add observability for attribution (#54039)
This adds traces, logs and metrics via the Operation abstraction. It is
my first time using it, but I believe I am using it correctly. To use
this we introduce a constructor which initializes the Operation for the
attribution service.

Test Plan: go test and sg start and visiting the prometheus metrics.