[fix] permission sync is now scheduled correctly for new users (#54258)
Previously we relied on users_with_no_perms query to schedule new users.
However this might not schedule anything for the new user, because the
user might already have permissions granted from pending permissions.

This resulted in sync job for the user never reaching the
permission_sync_jobs table in the database. Which in turn also resulted
in user not getting the permissions that they needed as fast as
possible. It also resutled in scheduled permission syncs not being
scheduled at all for the user, since the schedule depends on a record in
permission_sync_jobs table for each user.



## Test plan

unit tests