search: match lang specific files everywhere (#54350)
This fixes a bug where we only match certain files, such as Dockerfile
or BUILD.bazel, at root level when the user searches via the "lang:"
filter EG "r:sourcegraph$ lang:bazel"

Background:
For some languages, linguist, and consequently go-enry, specifies files
that don't match the language's extensions.

For example, for Starlark (alias of bazel), linguist lists the
extensions .bzl and .star and a couple of special files such as
`BUILD.bazel`. Note that ".bazel" is not listed as extension of
Starlark.

However, our regexp pattern for the special files matches only files at
root level.

This changes the regexp to be more permissive.

Cons:
- On the search page we now highlight a leading "/" for the new matches
which is suboptimal.