cody-gateway: hardcode notification thresholds per actor source (#54388)
As App usage of Gateway ramps up, we are expecting a _lot_ of actors to
hit their rate limits. This is already causing a lot of noise in the
channel. This change retains the previous notification thresholds for
product subscriptions, but disables them entirely for other actor
sources for now.

Env-based config is probably too brittle for this granularity so I think
we should just hardcode it.

Thread:
https://sourcegraph.slack.com/archives/C05BWKQB0R1/p1687904578368609

## Test plan

Unit tests

---------

Co-authored-by: Joe Chen <joe@sourcegraph.com>