executors/multi-queue: implement queue selection fairness (#53013)
- Closes https://github.com/sourcegraph/sourcegraph/issues/50642
- Depends on https://github.com/sourcegraph/sourcegraph/pull/53012

This PR adds a dequeue limit for each queue with a sliding window in
which the amount of dequeues are evaluated. The dequeues are stored in a
Redis cache.
Once the final list of candidate queues is established, a queue is
chosen randomly according to the configured weights.

### Unknowns
All of these are currently arbitarily picked:
- ~What should the window size be? Currently set to 5 minutes.~ ✅
configurable as env var
- ~What should the dequeue limits in a window be? Currently set to 50
for `batches`, 250 for `codeintel`.~ ✅ configurable in site config
- ~What should the weights for the queues be? Currently set to 4 for
`batches`, 1 for `codeintel` (`batches` should probably be
prioritised).~ ✅ configurable in site config
- ~Where should the janitor job be registered to clean up stale cache
entries? The routine should have access to the `*rcache.Cache`, unless
there's some way of getting an existing cache that I'm not aware of?~

### TODO
- ~Make the parameters configurable. Perhaps keep it undocumented for
now.~ ✅ Configurable as env vars and site config

### Dequeue flow chart

![243180731-a04773f0-346f-4463-9877-d95c7e2053cc](https://github.com/sourcegraph/sourcegraph/assets/2979513/f7bad466-6752-47e5-899a-9ae48ab53918)

## Test plan
Unit tests

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Randell Callahan <piszmogcode@gmail.com>