Tracing: remove LazyPrintf (#54564)
This removes the `LazyPrintf` method on `trace.Trace`. This method
existed because of `x/net/trace`, but now we can just use `AddEvent`
instead. It's not quite as ergonomic, but it provides more readable span
events (they're not named `LazyPrintf`) and helps discourage adding
massive payloads to a span.

This is another step towards standardization on OpenTelemetry.