sveltekit-prototype: Add separator component and setup storybook (#54674)
This commit adds a simple separator component to be used for resizing
panels (e.g. sidebars) and is used on file tree pages. The way the 
separator works is quite simple atm: It computes its position in
relation
to the size of the parent element and saves that value in the provided
store.
The parent component can use this value however it sees fit.
The current value is persisted in local storage.
I assume that this functionality will be expanded or changed in the
future
but it's a start.

I also setup storybook 7, following the default setup instructions. I
left everybody as default except for deleting the example stories it
came with. We can always change the structure later.
Initially I wasn't able to run `pnpm run storybook`, it was throwing an
error about not being able to import `ts-dedent`. Adding it to the root
`package.json` file seems to resolve this issue.

Repo page: 


https://github.com/sourcegraph/sourcegraph/assets/179026/fb6ba082-8959-47ed-b2ae-8e1aa351b64c

Storybook: 

![2023-07-06_14-34](https://github.com/sourcegraph/sourcegraph/assets/179026/558fdbfc-cbc6-45c3-973b-92d1931f1395)



## Test plan

Open repo page and use separator.

Run storybooks with `pnpm run storybooks`

Run existing storybooks in the workspace roots. It's not impacted by
adding a new version.