Switch to uncached doer in cody gateway (#54805)
Not only is the cache hit rate going to be really low and requests
large-ish, but also do we not want to retain any data. The dotcom
interactions seem to use the client from the gqlgen, so it doesn't look
like this needs to be changed (although, arguably consistency would be
nice).


## Test plan

Will need a manual test.