gitserver: introduce DB argument to NewClient (#54788)
To colocate forks and parent repositories on the same shard we will need
to consult the database when doing AddrForRepo. The is a partial
migration to injecting the DB into the gitserver client for that use
case. This PR does not actually use the field yet since we won't be able
to use it until all call sites are migrated. There are individual
commits containing mostly mechanical refactors to pass in a DB. There
are some other commits which manually migrated our use of NewClient. But
there are still many call sites which we didn't look at yet / do not
have a mechanical transformation. Those will be tackled in follow up
commits. For now all non-migrated sites now use the telling constructor
name NewClientDeprecatedNeedsDB.

Alternatively we considered introducing DB as an argument to
AddrForRepo, but this similarly will be a large change. We might still
do that since ClientSource is a singleton in practice. Additionally the
client having this dependency feels like how we would design it when
creating this from scratch.

Test Plan: go test. All changes which introduced the use of a DB used an
existing one. The only risk here is we introduce a new way to access the
db => something is nil along the way and there is no test coverage of
it. We only did that in one or two places, and they are safe. The vast
majority are mechanical changes which do not exercise anything new.

Co-authored-by: @indradhanush