embeddings: get rev from index instead of from job (#54919)
Right now we rely on the revision stored in `repo_embeddings_jobs` table
for incremental updates. This hasn't led to problems yet, but it might.

For example, if the jobs table doesn't reflect the state of the store
anymore, the diffs will be off (-> subtle bug) and we have to trigger a
complete reindex to fix this.

In general, we shouldn't consider the job table as authorative when it
comes to index metadata. Instead we should rely on the store or at least
a db table that is eventually consistent with the store.

Note:
While I was anyway touching the code, I simplifed the logic around
incremental updates a bit.

## Test plan
- CI, because this is mostly a refactor
- Manual testing: checked locally that incremental updates still work as
expected