sveltekit-prototype: Add initial version of expandable and accessible file tree (#54714)
This commit adds a reusable and accessible single select tree view, and
is used in the repo page to render the file tree sidebar.

Overview:

`TreeView.svelte` and `TreeNode.svelte` provide the generic tree
implementation. They depend on `TreeState`, which needs to be provided
as context by the caller, and `TreeProvider` which needs to be passed as
prop.

`FileTree.svelte` uses those components to provide the sidebar file
tree.

The code layout data loader fetches the file data for the current path
(or the "highest" path visited for the current directory, similar to how
the current web app works), which the code layout component uses to
create a `FileTreeProvider`.

Noteworthy:

- Tree state (focused, selected, expanded nodes) is stored globally (per
repo) so that we can restore the state when opening/closing the sidebar
or navigating back to the repo.
- I set the maximum number of fetched and shown nodes to 1000. The more
nodes we show the longer it takes for the (sub-)tree to be visible (we
have that problem in the current web app too) and I see no benefit from
showing 2500 instead of 1000 nodes (but we can always tweak this).
- Update performance (selecting nodes, change focused node) is really
good (irrespective of the number of rendered nodes (see video, compare
with current web app).
- I tried to make it fully keyboard accessible. I haven't implemented
some of the more "exotic" `aria-*` props like `aria-posinset`.
- Renders info node when not all nodes have been fetched/rendered (like
in the current web app)

What it doesn't do (yet):

- It doesn't handle single directory children. Supporting this would
require changes to data loading and representation.
- It doesn't render submodules.  
- Multiselection
- Unit tests



https://github.com/sourcegraph/sourcegraph/assets/179026/158dd428-9463-4098-bf68-fff4897e899a



## Test plan

Open repository, select a file and use arrow keys to navigate the file
tree.