search-blitz: remove group (#55065)
We have only ever had one group. I think this was introduced at a time
when we thought we may have different groups of queries, but we ended up
just relying on the query name for that.

Group was only used in local observability, but not in the metrics we
record in the frontend. So this _might_ break the local grafana
dashboard (which we can fix). But my main motivation was removing some
of the log noise.

Test Plan: go test