gitserver: Use address lookup cache conditionally  (#55188)
Part of #54741 and follow up on #55001.

## What 

In this commit we make two changes:

1. Add a new site config
`experimentalFeatures.useGitserverAddressLookupCache`
2. Use the site config to bypass the cache if it is explicitly set to
`false

## Why

In #55001 we added DB reads to the repository address lookup code along
with a cache to support git deduplication in the future. We want to test
the impact on a production instance with and without the cache.