bzl: remove from upsteam zoekt repo (#55516)
Updates our reference to the zoekt repo which no longer has bazel. Adds
back in the patches we were previously using to ensure we're building
statically, as it was observed this affected indexing behaviour in
previous testing.

## Test plan

Tested with images built from this branch on a compose deployment, which
no observable issues.