sveltekit: Show readme on repo and tree pages (#55616)
This PR changes the implementation of the repo page to show the contents
of the readme file, if it exist. If not it will show the repository's
description.
The file tree listing was removed. IMO the file tree sidebar works well
suffices.

Tree pages now show a (for now simple) file table, followed by a readme,
if available. I'm not super happy with how the table and readme look
together but we can iterate on that.

Repo page:


![2023-08-07_17-12_1](https://github.com/sourcegraph/sourcegraph/assets/179026/a345f2f7-1dbd-4fea-9fb3-cf167db39383)

Repo page collapsed:

![2023-08-07_17-13](https://github.com/sourcegraph/sourcegraph/assets/179026/b7e035b8-b8b3-4164-aefd-a76b3ca7bdec)

Tree page:

![2023-08-07_17-12](https://github.com/sourcegraph/sourcegraph/assets/179026/171572b7-6065-43be-8ce0-28315436ce02)

I will try to add stories once #55542 is merged.

## Test plan

`pnpm dev`