sveltekit: Refactor graphql setup (#55685)
Until now the app used a mixture of depcreated `requestGraphQL` calls
and Apollo queries.

This commit switches everything over to Apollo and uses Apollo's
InMemory cache instead of using `memoize` or `memoizedObservable`. This
brings a lot more consistency and predictability to the code.

A lot of the data we receive from the server is (or should be*)
immutable, since most of it is keyed to a specific revision. I updated
the queries to make sure that we use commit OIDs wherever possible.

One exception is the query for resolving the repo revision from the URL.
A symbolic rev might refer to different commits over time. If the URL
contains a symbolic rev we will always fetch the latest information from
the server. And fetching user information such as settings probably has
to be adjusted too, but since there is no code in the app that can
change settings this is not an issue yet.

I also changed how "nice" file URLs are computed. The GraphQL API allows
us to pass an "inputRevspec" to the call so that the backend can return
"nice" URLs that match the requested rev in the current URL. This
doesn't work very well with the cache. Instead I'm using the canonical
URL (which is always the same for a specific commit) and update it on
the client side.

Since I haven't implemented a revision selector yet there might be other
cache issues that occur when switching between different revisions. I'll
address them as they come up.

NOTE: Persisting the cache to local storage also works well, but in the
React app we are using separate caches for logged in and logged out
state, which is something we cannot determine yet at that point in the
SvelteKit app, so I didn't include it (though we could create a separate
cache key for development... 🤔 )

*: I said "should be" because I'm not sure how settings impact which
data is returned from the server. On the other hand, many of the
"memoized" don't take settings into account either when computing a key,
so it's probably fine.

## Test plan

`pnpm dev` and watched the network tab when switching between pages
`pnpm vitest`
`pnpm storybook`