gitserver: Start cleaning up server struct (#55671)
This PR starts moving some things off of the server struct that don't
actually need it.

## What do we gain? 

In tests, we don't need to construct the entire server each time, of
which we have no clue what of the fields are actually required in a
test. This makes the individual functions much easier to test.

Also, the server struct with about 100 receiver methods smells like a
huge anti-pattern. It's very hard to reason about what this monstrosity
is all doing, or maybe also.. not doing. I'm hoping this is slowly
becoming clearer as the individual subcomponents of the server become
more obvious by such refactors.

As a side effect of some of these things becoming standalone routines vs
being part of Server, we get:
- Them being goroutines they are tracked in the background routines in
our site-admin area
- And I personally find the Main function already much easier to read
and reason about

Most of these changes are extremely mechanical, because we now pass down
more arguments to functions instead of having everything on the server.
That causes a bunch of added diff, but should be fairly easy to scroll
past.

## Test plan

Tests are still passing, awaiting CI for integration tests, and started
a fresh gitserver installation and ended up with repos properly cloned.