Update p4api for building p4-fusion (#55891)
* Update p4api for building p4-fusion

Perforce changed the `p4api.tzg` artifact without changing the version, so the file hash is outdated. Switch to using `p4api-glibc2.3-openssl1.0.2.tgz` because the file hash is published, and because it includes the `glibc` and `openSSL` versions in the file name.

* Increment `epoch` in p4-fusion.yaml

Build in CI resulted in this message and a 222 error:

> The production package repository already contains a package with this version: p4-fusion-1.12-r2.apk
>    -> Production repository file path: gs://package-repository/main/x86_64/p4-fusion-1.12-r2.apk
> Resolve this issue by incrementing the `epoch` field in the package's YAML file.

Followed the instructions and incremented `epoch`.