searchjobs: PoC of listing search jobs (#56309)
This ignores all arguments sent via graphql and always lists all search
jobs created by a user. This should be something we can build on.

From reading the graphql schema we likely need a way to establish which
users we are listing for, unless that is somehow part of the query
field. In particular I am thinking of admins looking at search jobs.

Test Plan: manually ran the below graphql query locally

  query {
    searchJobs {
      nodes {
        id
        query
      }
    }
  }