search: make it possible to mock any RuntimeClients in client (#56695)
The goal is to mock gitserver client in another commit. This commit contains a
refactorings to the search client to make this possible. It does this by
storing inside of the search client the runtime clients and moving the
creation of the gitserver client into that field.

Test Plan: CI is sufficient for this change.