search jobs: implement stats resolver (#56717)
Contrary to the original plan, the stats resolver returns the aggregate state of the repo-rev jobs, NOT the repo jobs. This is easier to implement, because we don't have to worry about aggregating first on the repo level and then the job level. Let's see how the user experience is based on these stats and improve it if needed.

Notes:
- I had to add a constructor for `searchJobsResolver`, otherwise it is too easy to forget setting a field.
- While converting the call sites to use the contstrutor I came across an (now) unused searchJobsConnectionResolver, which I deleted.

Test plan: 
- updated unit test