user onboarding: Add dynamic queries to user onboarding (#56685)
This PR introduces a new "search query" step type to the onboarding tour. This type takes a query template and a list of code snippets. When the step is rendered it will make an "exploratory" query for each code snippet to find one that returns results. That query is then used in the onboarding tour.

The query template currently supports four variables: $$userorg, $$userrepo, $$userlangand$$snippet`.

The PR also removes the old logic for determining the user's language, which will be supplied by the user in a different way.