JetBrains: fix bug for access token description (#56897)
Previously, the "Description" field got auto-populated with the name of
the JetBrains IDE but the name wasn't correctly persisted after the user
created the token. In the UI, it would say "No description".

This PR fixes the problem by passing the default description value as
the initial value to the appropriate `setState` React hook.