search: filematch specific columns in exhaustive CSV output (#56834)
This updates the format of our downloads to be more in line with the the CSVs
users can already download from the web app. See the comments we added for
different decisions we made. Originally we wanted to have the same columns,
but we don't believe the columns in our export CSV are that useful. So we are
taking this opportunity to experiment with something that may be more useful.

It also updates the log format as follows:
- repo name instead of repo id
- more descriptive column names

Test Plan: updated test

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>