search: use high default limits for exhaustive (#57085)
Currently we use our default limits which are tuned for interactive and
regular searching. This commit introduces significantly higher limits
for search jobs (exhaustive):

- 1,000,000 file match limit (per repo@rev).
- 1 hour timeout.
- Ignore site config MaxTimeoutSeconds.

This is done by introducing a new Protocol "Exhaustive". All uses of
protocol before were for tuning limits, so this is the perfect place to
use it. Additionally it allows us to remove the hacky Exhaustive boolean
field.

There was a bunch of duplicated logic around how we computed file match
limits. This commit additionally removes those and makes all calls use
MaxResults. This is a tiny noop refactor in this commit.

Test Plan: Added a unit test which helps ensure we have higher limits.
Otherwise normal go test for catching regressions.